Updated Layout.unit and App.unit tests
